Broke out all epoll specific calls to epoll.c in preparation for a poll(2) implementation that will be stored in poll.c
